% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ce}
\alias{ce}
\title{ce}
\format{An \code{data.frame} object
\describe{
  \item{yr}{year, \code{numeric}}
  \item{mo}{month, \code{numeric}}
  \item{wk}{week, \code{numeric}}
  \item{flt}{fleet, \code{numeric}}
  \item{catch}{catch, \code{numeric}}
  \item{effort}{effort, \code{numeric}}
  ...
}}
\description{
An example dataset with catch and effort by year, month, week and fleet based on North 
Atlantic Albacore, derived from the ICCAT multifan-CL assessment
}
\keyword{data}

