\name{data}
\alias{rsdl}\alias{ce}
\title{Example Datasets}
\description{
\itemize{
     \item{rsdl}{ A data.frame with, \code{year,name,obs,hat,residual,residualLag,qqx,qqy,qqHat}},
     \item{ce}{ A data.frame with, \code{yr,mo,wk,flt,catch,effort}}
}
	Datasets can be loaded by issuing the \code{data} command, e.g. \code{data(rsdl)}.
}
\references{
tRFMO.
}
\keyword{datasets}
