% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diags-2box.R
\name{diags.vpa2box}
\alias{diags.vpa2box}
\alias{diags.vpa2box-method}
\alias{diags.vpa2box,character-method}
\title{diags.vpa2box}
\usage{
diags.vpa2box(file, ...)
}
\arguments{
\item{file}{the results file}

\item{...}{any other arguments}
}
\value{
an \code{data.frame} object
}
\description{
Reads in the fits to the indices of abundance from VPA2Box
}
\examples{
\dontrun{
diags.vpa2box(file)
} 

}
