\docType{methods}
\name{fget}
\alias{fget}
\title{fget}
\usage{
fget(file, package = "diags")
}
\arguments{
  \item{file}{the name of the file with subdir in a pacakge
  to reference}

  \item{package}{directs you to where you have installed
  the package}
}
\value{
\code{char} with path and file
}
\description{
Creates a path to a file in a package subdirectory
}
\examples{
x=1
}

