% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.VPA2BoxV2.R
\name{read.vpa2box}
\alias{read.vpa2box}
\title{read.vpa2box}
\usage{
read.vpa2box(file, ...)
}
\arguments{
\item{file}{the control file}

\item{m}{\code{vector} for natural mortality}

\item{minage}{the age of the first age in the data, 1 by default,}

\item{retros}{read in retrospectives, TRUE by default,}

\item{printFiles}{print the names of the individual file read in, FALSE by default}
}
\value{
an \code{FLStock} object
}
\description{
Reads in the results from VPA2Box into an \code{FLStock} object. This has slots for the estimates 
of numbers and fishing mortality-at-age and the assummed values of natural mortality, weight and 
maturity-at.age.
}
\examples{
\dontrun{

} 

}
