% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readU-methods.R
\docType{methods}
\name{readCpue}
\alias{readCpue}
\alias{readCpue-method}
\alias{readCpue,character,character-method}
\alias{readCpue,character,missing-method}
\title{readCpue}
\usage{
readCpue(file, method, ...)
}
\arguments{
\item{file}{the name of the file or dir which the data are to be read from.}

\item{method}{stock assessment method type one of \code{c("aspic","adapt","bsp","mfcl","ss","sam","vpa")}}

\item{...}{any other parameters}
}
\value{
a \code{data.frame} with the CPUE series.
}
\description{
read catch per unit effort data from a variety of stock assessment text files.
}
\examples{
\dontrun{
   dirMy="/home/laurie/Desktop/gcode/gbyp-sam/tests/aspic/bet/2010/run3"
   udat=readCpue(,what="")   
   head(udat)
   }
}
