% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeU-methods.R
\docType{methods}
\name{writeCpue,data.frame,character,character-method}
\alias{writeCpue,data.frame,character,character-method}
\alias{writeCpue-method}
\title{writeCpue}
\usage{
\S4method{writeCpue}{data.frame,character,character}(object, file,
  method = c("bsp", "ss", "vpa"), ...)
}
\arguments{
\item{object}{the name of the file or dir which the data are to be read from.}

\item{file}{name of text file to save data to.}

\item{method}{stock assessment method type one of \code{c("bsp","ss","vpa")}}

\item{...}{any other parameters}
}
\description{
Writes catch per unit effort in a data frame to a variety of stock assessment text files.
}
\examples{
\dontrun{
   data(diagAspic)
   writeAspic(diagAspic,"aspoc.inp","aspic")
   }
}
